Significant improvements to higherKindedTypes.
Improvements basically come from treating GenericTypeParameters as both
a TypeParameter and a GenericType, which means they are also treated as
an interface type since GenericType extends InterfaceType.

Treating GenericTypeParameters as InterfaceTypes works by making it look
like an interface that declares no members of its own but inherits
members from its constraint if it has one.

There also are changes related to inferring GenericTypeParameters where
they are not fully erased in getBaseSignature because we need to be able
to infer `T<_T>` from something like `T<_T> extends Functor<_T, T>` in
order to capture the fact that it references itself in its constraint.
See class `InvalidFunctor2` and `expectErrorD1` in
`higherKindedTypesLift.ts` for an example of this.